Limit number of shipment requests
