+++ src/scaup/crud/proposals.py
@@ -2,46 +2,22 @@ from typing import List
 
 from fastapi import HTTPException, status
 from lims_utils.models import Paged, ProposalReference
-from sqlalchemy import func, insert, select
+from sqlalchemy import insert, select
 from sqlalchemy.exc import MultipleResultsFound
 
-from ..auth import GenericUser
 from ..models.inner_db.tables import Sample, SessionType, Shipment
 from ..models.samples import SublocationAssignment
 from ..models.shipments import ShipmentIn
-from ..utils.auth import is_admin
-from ..utils.config import Config
 from ..utils.crud import assign_dcg_to_sublocation
 from ..utils.database import inner_db
 from ..utils.external import update_shipment_statuses
 
 
-def create_shipment(
-    proposal_reference: ProposalReference,
-    params: ShipmentIn,
-    user: GenericUser | None = None,
-):
+def create_shipment(proposal_reference: ProposalReference, params: ShipmentIn):
     # Proposal existence is already checked by Microauth
     session_type_id = inner_db.session.execute(
         select(SessionType.id).filter(SessionType.name == params.sessionType)
     ).scalar_one()
-
-    existing_shipment_count = inner_db.session.scalar(
-        select(func.count(Shipment.id)).filter(
-            Shipment.proposalCode == proposal_reference.code,
-            Shipment.proposalNumber == proposal_reference.number,
-            Shipment.visitNumber == proposal_reference.visit_number,
-        )
-    )
-
-    if existing_shipment_count > (Config.db.max_shipments_per_session - 1) and not (
-        user and is_admin(user.permissions)
-    ):
-        raise HTTPException(
-            status.HTTP_403_FORBIDDEN,
-            f"Only {Config.db.max_shipments_per_session} sample collections are allowed per session. Contact"
-            " staff if you require more.",
-        )
 
     new_shipment = inner_db.session.scalar(
         insert(Shipment).returning(Shipment),
